+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -8,6 +8,7 @@ Release History
 * BREAKING CHANGE:: Update '--no-push' to a pure flag in 'acr build' command.
 * Add 'show' and 'update' commands under 'acr repository' group.
 * Add '--detail' flag for 'show-manifests' and 'show-tags' to show more detailed information.
+* Add '--image' parameter to support get build details or logs by an image.
 
 2.0.28
 ++++++